# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/routes": {
      "get": {
        "operationId": "PlanningController_listRoutes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createRoute",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Crea una ruta o una NUEVA VERSIÓN (la anterior del mismo nombre queda «Obsoleta»).",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/routes/{id}/retire": {
      "post": {
        "operationId": "PlanningController_retireRoute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips": {
      "get": {
        "operationId": "PlanningController_listTrips",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Viajes del alcance (paginación SQL, facetas por estado, prioridad, servicio y condición).",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createTrip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Planifica un viaje (RN-002 sin superposición; recursos elegibles; servicio vigente opcional).",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips/{id}": {
      "get": {
        "operationId": "PlanningController_getTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips/{id}/assign": {
      "post": {
        "operationId": "PlanningController_assignTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips/{id}/gate": {
      "get": {
        "operationId": "PlanningController_gate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Evalúa el gate CTRL-001 y explica cada requisito (qué falla, quién corrige, qué regla).",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips/{id}/{action}": {
      "post": {
        "operationId": "PlanningController_act",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "enable",
                "dispatch",
                "arrival",
                "close",
                "cancel",
                "interrupt",
                "reassign",
                "reschedule"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Acción de lifecycle: enable | dispatch | arrival | close | cancel | interrupt | reassign | reschedule.",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/trips/{id}/messages": {
      "get": {
        "operationId": "PlanningController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/services": {
      "get": {
        "operationId": "PlanningController_listServices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createService",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/services/{id}": {
      "patch": {
        "operationId": "PlanningController_updateService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/services/{id}/transition": {
      "post": {
        "operationId": "PlanningController_transitionService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "planning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Gestión de Transportes — API",
    "description": "DOC-E-BE · Contratos API definitivos (REST /api/v1). Errores application/problem+json (ADR-013). Autorización por permiso + alcance + política; el tenant se resuelve SIEMPRE desde la sesión.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReasonDto": {
        "type": "object",
        "properties": {}
      },
...
      "RouteDto": {
        "type": "object",
        "properties": {}
      },
      "TripDto": {
        "type": "object",
        "properties": {}
      },
      "AssignTripDto": {
        "type": "object",
        "properties": {}
      },
      "TripActionDto": {
        "type": "object",
        "properties": {}
      },
      "MessageDto": {
        "type": "object",
        "properties": {}
      },
      "ServiceDto": {
        "type": "object",
        "properties": {}
      },
      "ServiceTransitionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}